+++ qa/lua/run.lua
@@ -53,6 +53,7 @@ local suites = {
     "test_mp_dailies",
     "test_mp_quest_boundary",
     "test_mp_shilling_ui_policy",
+    "test_mp_emporium_purchase",
     "test_wt_passive_charge",
     "test_wt_bolt_staff_overcharge",
     "test_wt_longbow_zoom_probe",